% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgelist_to_tbl.R
\name{edgelist_to_tbl}
\alias{edgelist_to_tbl}
\title{Converts edgelist (as stored in data) to a tibble with colums ~from, ~to (as required by tidygraph).}
\usage{
edgelist_to_tbl(edgelist)
}
\arguments{
\item{edgelist}{- single element character vector of edges: abcd means there's an edge from a to b, and from c to d.}
}
\description{
Converts edgelist (as stored in data) to a tibble with colums ~from, ~to (as required by tidygraph).
}
