% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggliph.R
\name{ggliph}
\alias{ggliph}
\title{ggplot of a glyph, via tidygraph.}
\usage{
ggliph(glyph_graph)
}
\arguments{
\item{glyph_graph}{the glyph to plot.}
}
\value{
A gg object ready to accept + geom_node_\* and/or geom_edge_\*.
}
\description{
Equivalent to ggraph() tweaked to accept a glyph graph.
}
