% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyph_graph.R
\name{glyph_graph}
\alias{glyph_graph}
\title{Generates a glyph as a tidygraph object.}
\usage{
glyph_graph(edges)
}
\arguments{
\item{edges}{- a data.frame, or object coercable to, with the edges for the glyph.}
}
\value{
An (undirected) tidygraph containing those edges.
}
\description{
Generates a glyph as a tidygraph object.
}
