% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes.R
\docType{data}
\name{nodes}
\alias{nodes}
\title{Nodes for any glyph graph}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 4 columns.}
\usage{
nodes
}
\description{
Returns a tibble ready to enter a tidygraph object as nodes argument.
Nodes hold an index which is referenced in edge creation.
}
\keyword{datasets}
